Add @Min and @Max annotations. Change those annotations to return double instead of int. Change "1 or 0" fields to Boolean. Add missing @Info annotations

+++ src/main/java/me/bogeymanEST/cmdhelper/tagdef/Max.java
@@ -38,5 +38,5 @@ public @interface Max {
      *
      * @return The maximum value
      */
-    public int value();
+    public double value();
 }
